requirements.txt
streamlit
langchain-google-genai
langchain
python-dotenv
requests
beautifulsoup4
